add code skeleton for running pre_promote in AsynExec
